Update gdal-20.rb
This updates gdal to the current 2.1.0 release, removes the internal numpy requirement, instead relying on homebrew supplied numpy. It allows python support for both python (2.x) and python3 support also.

Does this need to be a new gdal-21.rb formula? I'm not sure how you want to handle that, Larry.

Folks should test this against this repo as I have only used it in my fork.