Modified bottle building process to automatically update the Formulae with the new bottles.
Currently builds both sierra and high sierra bottles and merges them into the formula.
After updating all the formula, it adds them into a single commit and pushes the commit to master using the new deploy key.
This means we get a single commit with all the changes bottles, but Travis skips the new commit so we only get a single build.
If the build fails at any point, the bottles aren't built and no extra commits are made to the repo.

Also updated gitignore to skip IntelliJ files.